+++ app/(modals)/activities/[id]/expenses/new.tsx
@@ -1,229 +1,27 @@
-import { ActivityContext } from "@/context/ActivityContext";
-import { Expense } from "@/model/Expense";
-import { useLocalSearchParams } from "expo-router";
-import { useContext, useState } from "react";
-import {
-  ScrollView,
-  View,
-  TextInput,
-  Text,
-  StyleSheet,
-  Platform,
-} from "react-native";
+import NewExpense from "@/components/expenses/NewExpense";
 import Colors from "@/constant/Color";
-import {
-  PaperProvider,
-  Button,
-  Menu,
-  Checkbox,
-  List,
-  MD3LightTheme,
-  Divider,
-} from "react-native-paper";
+import { ActivityContext } from "@/context/ActivityContext";
+import { AntDesign } from "@expo/vector-icons";
+import { useLocalSearchParams, useNavigation } from "expo-router";
+import { useContext, useEffect } from "react";
+import { TouchableOpacity } from "react-native";
 
-export default function NewExpense() {
-  const { id } = useLocalSearchParams();
+export default function NewExpenseMoal() {
   const { get } = useContext(ActivityContext);
-
+  const { id } = useLocalSearchParams();
   const activity = get(id as string);
 
-  const [newExpense, setNewExpense] = useState({
-    description: "",
-    amount: "",
-    date: new Date().toISOString(),
-    paidBy: undefined,
-  });
+  const navigation = useNavigation();
 
-  const [newExpenseFor, setNewExpenseFor] = useState(
-    activity.participants.map((p) => p.id)
-  );
-
-  const [menuVisible, setMenuVisible] = useState(false);
-  console.log("Menu Visible:", menuVisible);
-
-  function onPaidByChange(v: any) {
-    console.log("heer");
-    setNewExpense({ ...newExpense, paidBy: v.id });
-    setMenuVisible(false);
-  }
-
-  function onDescriptionChange(description: string) {
-    setNewExpense({ ...newExpense, description });
-  }
-
-  function onAmountChange(amount: string) {
-    setNewExpense({ ...newExpense, amount });
-  }
-
-  function onPaidForChange(value: string) {
-    if (newExpenseFor.includes(value)) {
-      setNewExpenseFor((prev) => prev.filter((v) => v !== value));
-    } else {
-      setNewExpenseFor((prev) => [...prev, value]);
-    }
-  }
-
-  function onPaidForEveryone() {
-    setNewExpenseFor(activity.participants.map((p) => p.id));
-  }
-
-  function validated() {
-    if (!newExpense.description) {
-      alert("Expense need description");
-      return false;
-    }
-
-    if (!newExpense.amount || !parseFloat(newExpense.amount)) {
-      alert("Expense need amount");
-      return false;
-    }
-
-    if (!newExpense.paidBy) {
-      alert("Select who paid for this expense");
-      return false;
-    }
-
-    if (newExpenseFor.length === 0) {
-      alert("Expense has to be paid for at least one person");
-      return false;
-    }
-  }
-
-  function onSubmit() {
-    if (!validated()) {
-      return;
-    }
-    const expData = new Expense({
-      ...newExpense,
-      activityId: id,
-      date: new Date(newExpense.date),
+  useEffect(() => {
+    navigation.setOptions({
+      headerRight: () => (
+        <TouchableOpacity onPress={navigation.goBack}>
+          <AntDesign name="close" color={Colors.Card} size={24}></AntDesign>
+        </TouchableOpacity>
+      ),
     });
-    console.log(expData);
-    console.log(newExpenseFor);
-    return;
-    activity.addExpense(expData, newExpenseFor);
-  }
+  }, []);
 
-  return (
-    <PaperProvider>
-      <ScrollView style={styles.container}>
-        <Text style={styles.label}>Description</Text>
-        <TextInput
-          value={newExpense.description}
-          onChangeText={onDescriptionChange}
-          placeholder="What was the expense?"
-          style={styles.input}
-        />
-
-        <Text style={styles.label}>Amount</Text>
-        <TextInput
-          value={newExpense.amount}
-          onChangeText={onAmountChange}
-          keyboardType="numeric"
-          placeholder="e.g. 45.00"
-          style={styles.input}
-        />
-
-        <Text style={styles.label}>Paid By</Text>
-        <Menu
-          visible={menuVisible}
-          onDismiss={() => setMenuVisible(false)}
-          style={{ borderColor: Colors.SubText }}
-          anchor={
-            <Button
-              mode="elevated"
-              onPress={() => setMenuVisible(true)}
-              style={styles.dropdownButton}
-              textColor={Colors.Background}
-            >
-              {activity.participants.find((p) => p.id === newExpense.paidBy)
-                ?.name || "Select Payer"}
-            </Button>
-          }
-        >
-          {activity.participants.map((p) => (
-            <Menu.Item
-              key={p.name}
-              onPress={() => onPaidByChange(p)}
-              title={p.name}
-              style={{ width: 500 }}
-            />
-          ))}
-        </Menu>
-        <Text style={styles.label}>Paid For</Text>
-        <View style={{ marginBottom: 24 }}>
-          {newExpenseFor.length < activity.participants.length && (
-            <>
-              <Checkbox.Item
-                label="Every One"
-                status={
-                  newExpenseFor.length === activity.participants.length
-                    ? "checked"
-                    : "unchecked"
-                }
-                onPress={onPaidForEveryone}
-                mode="android"
-              ></Checkbox.Item>
-              <Divider></Divider>
-            </>
-          )}
-
-          {activity.participants.map((participant) => (
-            <Checkbox.Item
-              key={participant.id}
-              label={participant.name}
-              status={
-                newExpenseFor.includes(participant.id) ? "checked" : "unchecked"
-              }
-              onPress={() => onPaidForChange(participant.id)}
-              mode="android" // or 'ios' for a different look
-            />
-          ))}
-        </View>
-
-        <Button
-          mode="contained"
-          onPress={onSubmit}
-          style={styles.submitButton}
-          labelStyle={styles.submitButtonLabel}
-        >
-          Add Expense
-        </Button>
-      </ScrollView>
-    </PaperProvider>
-  );
+  return <NewExpense activity={activity}></NewExpense>;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 16,
-    gap: 16,
-  },
-  label: {
-    fontSize: 14,
-    marginBottom: 4,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: Colors.SubText,
-    borderRadius: 8,
-    paddingHorizontal: 12,
-    paddingVertical: Platform.OS === "ios" ? 12 : 8,
-    marginBottom: 16,
-  },
-  dropdownButton: {
-    borderRadius: 8,
-    marginBottom: 16,
-    backgroundColor: Colors.Primary,
-  },
-  submitButton: {
-    borderRadius: 8,
-    elevation: 2, // raised shadow
-    backgroundColor: Colors.Primary,
-  },
-  submitButtonLabel: {
-    color: Colors.Background,
-    fontWeight: "bold",
-  },
-});
